Move broker ns creation to apis
....so that submariner-addon can use it.
